webapi/webapi-contactsmanager-sysapps-tests/contactsmanager/ContactAddress_attributes.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, Kang <kangx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>ContactsManager Test: ContactAddress_attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/2012/sysapps/contacts-manager-api/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>
var contactAddress;

setup(function () {
  contactAddress = new ContactAddress({
    types: ['work'],
    preferred: true,
    streetAddress: 'street',
    locality: 'local',
    region: 'region',
    postalCode: 'postal',
    countryName: 'china'
  });
});

test(function () {
  assert_true("types" in contactAddress, "The ContactAddress.types");
  assert_equals(typeof contactAddress.types, "object", "The type of ContactAddress.types");
}, "Check if the attribute ContactAddress.types exists and typeof object");

test(function () {
  assert_true("preferred" in contactAddress, "The ContactAddress.preferred");
  assert_equals(typeof contactAddress.preferred, "boolean", "The type of ContactAddress.preferred");
}, "Check if the attribute ContactAddress.preferred exists and typeof boolean");

test(function () {
  assert_true("streetAddress" in contactAddress, "The ContactAddress.streetAddress");
  assert_equals(typeof contactAddress.streetAddress, "string", "The type of ContactAddress.streetAddress");
}, "Check if the attribute ContactAddress.streetAddress exists and typeof string");

test(function () {
  assert_true("locality" in contactAddress, "The ContactAddress.locality");
  assert_equals(typeof contactAddress.locality, "string", "The type of ContactAddress.locality");
}, "Check if the attribute ContactAddress.locality exists and typeof string");

test(function () {
  assert_true("region" in contactAddress, "The ContactAddress.region");
  assert_equals(typeof contactAddress.region, "string", "The type of ContactAddress.region");
}, "Check if the attribute ContactAddress.region exists and typeof string");

test(function () {
  assert_true("postalCode" in contactAddress, "The ContactAddress.postalCode");
  assert_equals(typeof contactAddress.postalCode, "string", "The type of ContactAddress.postalCode");
}, "Check if the attribute ContactAddress.postalCode exists and typeof string");

test(function () {
  assert_true("countryName" in contactAddress, "The ContactAddress.countryName");
  assert_equals(typeof contactAddress.countryName, "string", "The type of ContactAddress.countryName");
}, "Check if the attribute ContactAddress.countryName exists and typeof string");
</script>
